+++ src/main/layer/lstm.py
@@ -1,197 +1,134 @@
 import numpy as np
-import pandas as pd
-from matplotlib import pyplot as plt
-
-from Sequential import Sequential
-from binary_cross_entropy import BinaryCrossEntropy
-from dense import Dense
-from layer import Layer
-from sigmoid import Sigmoid
-from tanh import Tanh
-from text_vetorization import TextVectorization
 
 
-class LSTMLayer(Layer):
-    input = None
-    input_shape = None
-    output_shape = None
+def sigmoid(x):
+    np.clip(x, -500, 500, out=x)
+    return 1 / (1 + np.exp(-x))
 
-    def sigmoid(self, x):
-        x = x.clip(-700, 700)
-        return 1 / (1 + np.exp(-x))
 
-    def sigmoid_derivative(self, x):
-        return x * (1 - x)
+class LSTM:
+    def __init__(self, unit, input_shape, return_sequences=False):
+        self.unit = unit
+        self.time, self.feature = input_shape
+        self.return_sequences = return_sequences
 
-    def tanh_derivative(self, x):
-        return 1 - (np.tanh(x)) ** 2
+        self.Wa = np.random.randn(self.feature, unit)
+        self.Wi = np.random.randn(self.feature, unit)
+        self.Wf = np.random.randn(self.feature, unit)
+        self.Wo = np.random.randn(self.feature, unit)
 
-    def tanh(self, x):
-        return np.tanh(x)
+        self.Ua = np.random.randn(unit, unit)
+        self.Ui = np.random.randn(unit, unit)
+        self.Uf = np.random.randn(unit, unit)
+        self.Uo = np.random.randn(unit, unit)
 
-    def __init__(self, units, input_shape=None, return_sequence=False):
-        if input_shape is not None:
-            self.input_shape = input_shape
-        self.units = units
-        self.return_sequence = return_sequence
+        self.Ba = np.zeros((1, unit))
+        self.Bi = np.zeros((1, unit))
+        self.Bf = np.zeros((1, unit))
+        self.Bo = np.zeros((1, unit))
 
-    def _forward(self, x):
-        self.batch_size, self.time_step, self.features = x.shape
+    def forward(self, x):
+        self.input = x
+        self.batch = x.shape[0]
 
-        self.output = np.zeros((self.batch_size, self.time_step, self.units))
-        self.state = np.zeros((self.batch_size, self.time_step, self.units))
+        self.a = np.zeros((self.time, self.batch, self.unit))
+        self.i = np.zeros((self.time, self.batch, self.unit))
+        self.f = np.zeros((self.time, self.batch, self.unit))
+        self.o = np.zeros((self.time, self.batch, self.unit))
 
-        self.f = np.zeros((self.batch_size, self.time_step, self.units))
-        self.i = np.zeros((self.batch_size, self.time_step, self.units))
-        self.o = np.zeros((self.batch_size, self.time_step, self.units))
-        self.a = np.zeros((self.batch_size, self.time_step, self.units))
+        self.state = np.zeros((self.time, self.batch, self.unit))
+        self.output = np.zeros((self.time, self.batch, self.unit))
 
-        self.X = np.zeros((self.batch_size, self.time_step, self.features + self.units))
-        self.x = x
+        output_prev = np.zeros((self.batch, self.unit))
+        state_prev = np.zeros((self.batch, self.unit))
 
-        prev_state = np.zeros((self.batch_size, self.units))
-        prev_output = np.zeros((self.batch_size, self.units))
+        for t in range(self.time):
+            self.a[t] = np.tanh((x[:, t] @ self.Wa) + (output_prev @ self.Ua) + self.Ba)
+            self.i[t] = sigmoid((x[:, t] @ self.Wi) + (output_prev @ self.Ui) + self.Bi)
+            self.f[t] = sigmoid((x[:, t] @ self.Wf) + (output_prev @ self.Uf) + self.Bf)
+            self.o[t] = sigmoid((x[:, t] @ self.Wo) + (output_prev @ self.Uo) + self.Bo)
 
-        for t in range(self.time_step):
-            x_t = np.column_stack((prev_output, x[:, t, :]))
-            self.X[:, t, :] = x_t
+            self.state[t] = (self.a[t] * self.i[t]) + (self.f[t] * state_prev)
+            self.output[t] = np.tanh(self.state[t]) * self.o[t]
 
-            self.a[:, t] = self.tanh(np.dot(x_t, self.Wa) + self.Ba)
-            self.i[:, t] = self.sigmoid(np.dot(x_t, self.Wi) + self.Bi)
-            self.f[:, t] = self.sigmoid(np.dot(x_t, self.Wf) + self.Bf)
-            self.o[:, t] = self.sigmoid(np.dot(x_t, self.Wo) + self.Bo)
+            state_prev = self.state[t]
+            output_prev = self.output[t]
 
-            self.state[:, t, :] = (self.f[:, t] * prev_state) + (self.i[:, t, :] * self.a[:, t])
-            self.output[:, t, :] = self.o[:, t] * self.tanh(self.state[:, t])
+        if self.return_sequences:
+            return np.moveaxis(self.output, 0, 1)
+        return self.output[-1]
 
-            prev_state = self.state[:, t, :]
-            prev_output = self.output[:, t, :]
+    def backward(self, d_loss, learning_rate):
+        dWa = np.zeros((self.feature, self.unit))
+        dWi = np.zeros((self.feature, self.unit))
+        dWf = np.zeros((self.feature, self.unit))
+        dWo = np.zeros((self.feature, self.unit))
 
-        if self.return_sequence:
-            return self.output
-        return self.output.mean(axis=1)
+        dUa = np.zeros((self.unit, self.unit))
+        dUi = np.zeros((self.unit, self.unit))
+        dUf = np.zeros((self.unit, self.unit))
+        dUo = np.zeros((self.unit, self.unit))
 
-    def _backward(self, d_loss, learning_rate):
-        delta_output_next = np.zeros((self.batch_size, self.units))
-        delta_state_next = np.zeros((self.batch_size, self.units))
+        dBa = np.zeros((1, self.unit))
+        dBi = np.zeros((1, self.unit))
+        dBf = np.zeros((1, self.unit))
+        dBo = np.zeros((1, self.unit))
 
-        dWa = np.zeros_like(self.Wa)
-        dWi = np.zeros_like(self.Wi)
-        dWf = np.zeros_like(self.Wf)
-        dWo = np.zeros_like(self.Wo)
+        d_output_next = np.zeros((self.batch, self.unit))
+        d_state_next = np.zeros((self.batch, self.unit))
+        next_f = np.zeros((self.batch, self.unit))
 
-        dBa = np.zeros_like(self.Ba)
-        dBi = np.zeros_like(self.Bi)
-        dBf = np.zeros_like(self.Bf)
-        dBo = np.zeros_like(self.Bo)
-
-        for t in reversed(range(self.time_step)):
-            prev_state = self.state[:, t, :]
+        d_output = d_loss
+        for t in reversed(range(self.time)):
             if t == 0:
-                prev_state = np.zeros_like(prev_state)
+                output_prev = np.zeros((self.batch, self.unit))
+                state_prev = np.zeros((self.batch, self.unit))
+            else:
+                output_prev = self.output[t - 1]
+                state_prev = self.state[t - 1]
+            # Return Seq False (32,n)
+            # Return Seq True  (32,32,n)
+            d_out = d_output + d_output_next
+            d_state = d_out * self.o[t] * (1 - np.tanh(self.state[t]) ** 2) + (d_state_next * next_f)
 
-            delta_out = d_loss + delta_output_next
+            d_a = d_state * self.i[t] * (1 - self.a[t] ** 2)
+            d_i = d_state * self.a[t] * self.i[t] * (1 - self.i[t])
+            d_f = d_state * state_prev * self.f[t] * (1 - self.f[t])
+            d_o = d_out * np.tanh(self.state[t]) * self.o[t] * (1 - self.o[t])
 
-            delta_state = delta_out * self.o[:, t, :] * self.tanh_derivative(self.state[:, t, :]) + delta_state_next
-            delta_a = delta_state * self.i[:, t, :] * self.tanh_derivative(self.a[:, t, :])
-            delta_i = delta_state * self.a[:, t, :] * self.sigmoid_derivative(self.i[:, t, :])
-            delta_f = delta_state * prev_state * self.sigmoid_derivative(self.f[:, t, :])
-            delta_o = delta_out * self.tanh(self.state[:, t, :]) * self.sigmoid_derivative(self.o[:, t, :])
+            dWa += (d_a.T @ self.input[:, t]).T
+            dWi += (d_i.T @ self.input[:, t]).T
+            dWf += (d_f.T @ self.input[:, t]).T
+            dWo += (d_o.T @ self.input[:, t]).T
 
-            dWa -= self.X[:, t].T.dot(delta_a)
-            dWi -= self.X[:, t].T.dot(delta_i)
-            dWf -= self.X[:, t].T.dot(delta_f)
-            dWo -= self.X[:, t].T.dot(delta_o)
+            dUa += np.sum(d_a * output_prev, axis=0, keepdims=True).T
+            dUi += np.sum(d_i * output_prev, axis=0, keepdims=True).T
+            dUf += np.sum(d_f * output_prev, axis=0, keepdims=True).T
+            dUo += np.sum(d_o * output_prev, axis=0, keepdims=True).T
 
-            dBa -= delta_a.mean(axis=0, keepdims=True)
-            dBi -= delta_i.mean(axis=0, keepdims=True)
-            dBf -= delta_f.mean(axis=0, keepdims=True)
-            dBo -= delta_o.mean(axis=0, keepdims=True)
+            dBa += np.sum(d_a, axis=0)
+            dBi += np.sum(d_i, axis=0)
+            dBf += np.sum(d_f, axis=0)
+            dBo += np.sum(d_o, axis=0)
 
-            dX = np.dot(delta_a , self.Wa.T) + np.dot(delta_i , self.Wi.T) + np.dot(delta_f , self.Wf.T) + np.dot(delta_o , self.Wo.T)
+            d_output = np.zeros_like(d_output)
+            next_f = self.f[t]
+            d_state_next = d_state
+            d_output_next = (d_a @ self.Ua) + (d_i @ self.Ui) + (d_f @ self.Uf) + (d_o @ self.Uo)
 
-            delta_output_next = dX[:, :self.units]
-            delta_state_next = delta_state * self.f[:, t, :]
+        self.Wa -= dWa * learning_rate
+        self.Wi -= dWi * learning_rate
+        self.Wf -= dWf * learning_rate
+        self.Wo -= dWo * learning_rate
 
-        clip = lambda x: np.nan_to_num(np.sum(x / self.batch_size, axis=0, keepdims=True))
+        self.Ua -= dUa * learning_rate
+        self.Ui -= dUi * learning_rate
+        self.Uf -= dUf * learning_rate
+        self.Uo -= dUo * learning_rate
 
-        self.Wa += learning_rate * clip(dWa)
-        self.Wi += learning_rate * clip(dWi)
-        self.Wf += learning_rate * clip(dWf)
-        self.Wo += learning_rate * clip(dWo)
+        self.Ba -= dBa * learning_rate
+        self.Bi -= dBi * learning_rate
+        self.Bf -= dBf * learning_rate
+        self.Bo -= dBo * learning_rate
 
-        self.Ba += learning_rate * dBa
-        self.Bi += learning_rate * dBi
-        self.Bf += learning_rate * dBf
-        self.Bo += learning_rate * dBo
-
-        return delta_output_next
-
-    def compile(self, input_shape=None):
-        if input_shape is not None:
-            self.input_shape = input_shape
-
-        self.time_step, self.features = self.input_shape
-
-        if self.return_sequence:
-            self.output_shape = (self.time_step, self.units,)
-        else:
-            self.output_shape = (self.units,)
-
-        self.Wf = np.random.randn(self.features + self.units, self.units)
-        self.Wi = np.random.randn(self.features + self.units, self.units)
-        self.Wa = np.random.randn(self.features + self.units, self.units)
-        self.Wo = np.random.randn(self.features + self.units, self.units)
-
-        self.Bf = np.zeros(shape=(1, self.units))
-        self.Bi = np.zeros(shape=(1, self.units))
-        self.Ba = np.zeros(shape=(1, self.units))
-        self.Bo = np.zeros(shape=(1, self.units))
-
-        print(f"LSTM : {self.input_shape} -> {self.output_shape}")
-
-
-if __name__ == '__main__':
-    time_step = 3
-    feature = 1
-
-    layers = [
-        LSTMLayer(32, input_shape=(time_step, feature), return_sequence=False),
-        Dense(8),
-        Sigmoid()
-    ]
-
-    for i, layer in enumerate(layers):
-        if i != 0:
-            layer.compile(input_shape=layers[i - 1].output_shape)
-            continue
-        layer.compile()
-
-    loss = BinaryCrossEntropy()
-
-    errors = []
-
-    x = np.array([[[1], [2], [1]], [[1], [2], [4]], [[3], [4], [1]], [[4], [4], [1]]])
-    y = np.array([[1], [0], [0], [1]])
-
-    for i in range(10240):
-        p = x
-        for layer in layers:
-            p = layer.forward(p)
-
-        error = loss(p, y, derivative=True)
-
-        for layer in reversed(layers):
-            error = layer.backward(error, 0.01)
-
-        err = loss(p, y)
-
-        print(f'Epoch {i}, error={err:0.4}')
-        errors.append(err)
-
-    for layer in layers:
-        x = layer.forward(x)
-
-    plt.plot(errors, label='error')
-    plt.legend()
-    plt.show()
+        return d_output_next
